Move transaction column definitions to a new file
